requirements.txt
langgraph
langchain
langchain_google_genai
python-dotenv
ipython
gradio